pyproject.toml
[tool.poetry]
name = "djangoproject1"
version = "0.1.0"
description = ""
authors = ["Michal Nowak <m.nowak@bankier.pl >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
